Conflicts:
	trunk/src/protocol/messagefactory.cc
	trunk/src/protocol/objectdatamsg.cc
	trunk/src/protocol/putobjectendreply.cc
	trunk/src/protocol/putobjectendrequest.cc
	trunk/src/protocol/putobjectinitreply.cc
	trunk/src/protocol/putobjectinitrequest.cc
	trunk/src/protocol/putsegmentinitrequest.cc
	trunk/src/protocol/putsegmentinitrequest.hh
	trunk/src/protocol/segmentdatamsg.cc

+++ trunk/src/communicator/socket.cc
@@ -154,6 +154,6 @@ void Socket::set_non_blocking(const bool b) {
 
 }
 
-int Socket::getSockfd() {
-	return m_sock;
+uint32_t Socket::getSockfd() {
+	return (uint32_t)m_sock;
 }
